requirements.txt
APScheduler==3.9.1
beautifulsoup4==4.11.1
Django==4.1.1
djangorestframework==3.13.1
requests==2.28.1
selenium==4.4.3
webdriver-manager==3.8.3
lxml==4.9.1
python-dotenv==0.21.0